Move clique utility functions to clique module
