+++ strawberryfields/gbs/utils.py
@@ -36,9 +36,6 @@ Summary
     subgraph_adjacency
     validate_graph
     is_subgraph
-    is_clique
-    c_0
-    c_1
 
 Code details
 ^^^^^^^^^^^^
@@ -148,101 +145,3 @@ def is_subgraph(subgraph: Iterable, graph: nx.Graph):
         return set(subgraph).issubset(graph.nodes)
     except TypeError:
         raise TypeError("subgraph and graph.nodes must be iterable")
-
-
-def is_clique(graph: nx.Graph) -> bool:
-    """Determines if the input graph is a clique. A clique of :math:`n` nodes has exactly :math:`n(
-    n-1)/2` edges.
-
-    **Example usage:**
-
-    >>> graph = nx.complete_graph(10)
-    >>> is_clique(graph)
-    True
-
-    Args:
-        graph (nx.Graph): the input graph
-
-    Returns:
-        bool: ``True`` if input graph is a clique and ``False`` otherwise
-    """
-    edges = graph.edges
-    nodes = graph.order()
-
-    return len(edges) == nodes * (nodes - 1) / 2
-
-
-def c_0(clique: list, graph: nx.Graph):
-    """Generates the set :math:`C_0` of nodes that are connected to all nodes in the input
-    clique subgraph.
-
-    The set :math:`C_0` is defined in :cite:`pullan2006phased` and is used to determine nodes
-    that can be added to the current clique to grow it into a larger one.
-
-    **Example usage:**
-
-    >>> graph = nx.complete_graph(10)
-    >>> clique = [0, 1, 2, 3, 4]
-    >>> c_0(clique, graph)
-    [5, 6, 7, 8, 9]
-
-    Args:
-        clique (list[int]): a subgraph specified by a list of nodes; the subgraph must be a clique
-        graph (nx.Graph): the input graph
-
-    Returns:
-        list[int]: a list containing the :math:`C_0` nodes for the clique
-
-    """
-    if not is_clique(graph.subgraph(clique)):
-        raise ValueError("Input subgraph is not a clique")
-
-    clique = set(clique)
-    c_0_nodes = []
-    non_clique_nodes = set(graph.nodes) - clique
-
-    for i in non_clique_nodes:
-        if clique.issubset(graph.neighbors(i)):
-            c_0_nodes.append(i)
-
-    return c_0_nodes
-
-
-def c_1(clique: list, graph: nx.Graph):
-    """Generates the set :math:`C_1` of nodes that are connected to all but one of the nodes in
-    the input clique subgraph.
-
-    The set :math:`C_1` is defined in :cite:`pullan2006phased` and is used to determine outside
-    nodes that can be swapped with clique nodes to create a new clique.
-
-    **Example usage:**
-
-    >>> graph = nx.wheel_graph(5)
-    >>> clique = [0, 1, 2]
-    >>> c_1(clique, graph)
-    [(1, 3), (2, 4)]
-
-    Args:
-        clique (list[int]): a subgraph specified by a list of nodes; the subgraph must be a clique
-        graph (nx.Graph): the input graph
-
-    Returns:
-       list[tuple(int)]: A list of tuples. The first node in the tuple is the node in the clique
-       and the second node is the outside node it can be swapped with.
-   """
-    if not is_clique(graph.subgraph(clique)):
-        raise ValueError("Input subgraph is not a clique")
-
-    clique = set(clique)
-    c_1_nodes = []
-    non_clique_nodes = set(graph.nodes) - clique
-
-    for i in non_clique_nodes:
-        neighbors_in_subgraph = clique.intersection(graph.neighbors(i))
-
-        if len(neighbors_in_subgraph) == len(clique) - 1:
-            to_swap = clique - neighbors_in_subgraph
-            (i_clique,) = to_swap
-            c_1_nodes.append((i_clique, i))
-
-    return c_1_nodes
